pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager - web tool for task managment between co-workers."
authors = ["DREU007 <dreu007@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.6"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
django-extensions = "^3.2.3"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-bootstrap5 = "^23.3"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
django-filter = "^23.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
ipython = "^8.16.1"
pudb = "^2022.1.3"
coverage = "^7.3.2"
selenium = "^4.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
